requirement.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
langchain-community
faiss-cpu
langchain_google_genai